requirements.txt
fastapi~=0.70.0
packaging >= 20.9
pydantic~=1.8.2
SQLAlchemy~=1.4.21
databases~=0.5.3
Werkzeug~=2.0.1
aiosqlite~=0.17.0
python-multipart~=0.0.5
python-jose[cryptography]

uvicorn